rename teleop class, clean up
