+++ lerobot_robot_ros/lerobot_robot_ros/__init__.py
@@ -1,2 +1,2 @@
-from .config_ros import AnninAR4Config, ROS2Config, SO101ROSConfig
-from .ros import ROS2Robot
+from .config import AnninAR4Config, ROS2Config, SO101ROSConfig
+from .robot import SO101ROS, AnninAR4, ROS2Robot
