+++ lerobot_robot_ros/lerobot_robot_ros/ros_interface.py
@@ -28,7 +28,7 @@ from sensor_msgs.msg import JointState
 from std_msgs.msg import Float64MultiArray
 from trajectory_msgs.msg import JointTrajectory, JointTrajectoryPoint
 
-from .config_ros import ActionType, GripperActionType, ROS2InterfaceConfig
+from .config import ActionType, GripperActionType, ROS2InterfaceConfig
 from .moveit_servo import MoveIt2Servo
 
 logger = logging.getLogger(__name__)
